move world to separate module
