+++ Engine/Source/universe/src/lib.rs
@@ -2,22 +2,18 @@
 //!
 //! The **Universe** is `DirkEngine`'s ECS system.
 
-use std::{any::TypeId, collections::HashMap};
+use std::collections::HashMap;
 
-use crate::{
-    components::{AnyComponent, EntityComponents},
-    query::Query,
-    systems::{UniverseSystem, UniverseSystemHandle, UniverseSystemStorage},
-};
+use crate::systems::{UniverseSystem, UniverseSystemHandle, UniverseSystemStorage};
 
 pub mod components;
 pub mod query;
 pub mod systems;
+mod world;
+pub use world::{World, WorldId};
 
 /// A unique, opaque identifier for a spawned entity.
 pub type Entity = u32;
-/// An identifier that distinguishes multiple [`World`] instances from each other.
-pub type WorldId = u32;
 
 /// This struct is the manager for all the worlds.
 #[derive(Default)]
@@ -70,38 +66,5 @@ impl Universe {
     /// [`UniverseSystemHandle`].
     pub fn unregister_universe_system(&mut self, handle: UniverseSystemHandle) {
         self.universe_systems.remove(handle);
-    }
-}
-
-/// This is a world. It has entities and components.
-#[derive(Default)]
-pub struct World {
-    id: WorldId,
-    next_id: Entity,
-    alive: Vec<Entity>,
-    entity_components: EntityComponents,
-    /// These should only be [`components::WorldComponent`]. This is
-    /// guaranteed by the [`World`] API. Please make sure to add these
-    /// properly internally.
-    components: HashMap<TypeId, Box<dyn AnyComponent>>,
-}
-
-impl World {
-    /// Creates a new empty world with the specified ID.
-    #[must_use]
-    fn new(id: WorldId) -> Self {
-        Self {
-            id,
-            ..Self::default()
-        }
-    }
-    /// Calls all the destruction [`System`]s on the world
-    fn destroy(&mut self) {
-        todo!("call all the world systems for destruction")
-    }
-
-    #[must_use]
-    fn query(&self, query: &Query) -> Vec<Entity> {
-        todo!("Query for entities")
     }
 }
